+++ pkg-py/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+* Added rich tool UI support using shinychat development version and chatlas >= 0.11.1 (#67)
+
 ## [0.2.1] - 2025-09-04
 
 * Fixed an issue with the query tool when used with SQLAlchemy data sources. (@npelikan #79)